Add class get answer to find answer
This class is not working properly yet. There is a bug due to which
on jumping to every circle after some time, the jump overshoots the
required circle and also there was a segmentation fault. Trying
another approach now which includes drawing bounding boxes around
every circle and finding histogram in each row/column and estimating
answer.