Add feature for finding answer string
This feature is running properly in most of the cases but a bug
is observed.
Bug is that if there is no answer for a particular row
(as in A_11t15) but due to not 100% accurate grids, some part
of previous answer comes in answer below that making number of
black pixel value for that answer greater than 0, then the code
assumes it as the answer.