Update code
Trying to remove all individual variables used in main.py
for ech question and intead maikng the code such that
main.py is not necessary to tampered with when changing
between omr sheets. Currently there are some places which
needs to be tampered im main.py for the same.
Creating a new class for the same to combine different classes.