Add file altToCornerCircles.py
This file aims to find the alternative of the corner circles
with reference to which we do projective transform on the
image. It is seen that in most of the cases, corner circles
are not provided but instead we have horizontal small blocks
along both of the verticle edges of the OMR Sheet.