Add template matching feature and filter rctangles
Template matching is done to find the horizontal rectangles along
the verticle edges in the image. The template provided is hard
coded and extracted from the image from a particular position.
This needs to be changed at the end as the template will not be
present at the same place for every image.
Also the output rectangles from template matching were found to be
repeating or some rectangles denoted the same rectangle in the
image. These are also filtered and unnecessary rectangles are
deleted from list.