Refactor code for another omr
It is seen that for a different omr, all of the guiding
rectangles are not found for the template, but if the
template size is reduced and increased, all the guiding
rectangles are atleast found once. The code present now
\is to test for the same thing but is dirty. Code
cleanup will be applied and coordinates of all the rectangles
is to be found.