Add template images and refactor folder structure
It is not that the single template was not sufficient. For other
omr sheet, the boxes may be a little bit tilted or also of
different sizes. So it was considered necessary to add more
template image of different size and with some rotation. If
necessary, more template images might be added afterwards.