Add feature to shrink box wrt boundary.
A box is shrinked if all the pixels on the boundary of the box
are black and the central pixel of the box is not black.
This was considered necessary because some boxes after using
varying template, size of some boxes became so large that it
began intersecting the actual box above or below it due to
which few actual boxes were neglected and the output was not
coming correctly.