Add limiting conditions
This is added because if we encounter a contour at the boundary,
and there the white region exists, the code will try to increase
the coordinates but then the coordinates will overshoot the
image and in the next iteration it throws an error.
To avoid this, boundary conditions are added.