Add limiting cond. while expanding side
Limiting condition is added so that it doesnot throw runtime
error and also ExpandSideBy is converted to list which denotes
the different lengths of the 2 sides(top and bottom) by which
they will be expanded. Presviously they were expanded by same
lengths which was not necessary. Other code is also modified
accordingly.